# backend: improve ParseIntervalStringToTimeDuration method performance in gtime package by ~93%